type fix

+++ src/main/java/com/insurance/portal/java/repository/UserRepository.java
@@ -3,11 +3,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.insurance.portal.java.entity.*;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
 	User getUserByEmail(String email);
 
 	User getUserByPhone(Long phone);
-
 
 }
